Rename write/patterns => write/collections
