+++ src/write/collections.js
