+++ src/write/index.js
@@ -1,7 +1,11 @@
 import writePages from './pages';
+import writeCollections from './collections';
 
 function write (drizzleData) {
-  return writePages(drizzleData);
+  return Promise.all([
+    writePages(drizzleData),
+    writeCollections(drizzleData)
+  ]).then(() => drizzleData);
 }
 
 export default write;
